Move GetAssistancePage into workspace folder
