+++ src/libs/Navigation/AppNavigator/ModalStackNavigators.js
@@ -6,7 +6,7 @@ import NewChatPage from '../../../pages/NewChatPage';
 import NewGroupPage from '../../../pages/NewGroupPage';
 import SearchPage from '../../../pages/SearchPage';
 import DetailsPage from '../../../pages/DetailsPage';
-import GetAssistancePage from '../../../pages/GetAssistancePage';
+import GetAssistancePage from '../../../pages/workspace/GetAssistancePage';
 import IOURequestPage from '../../../pages/iou/IOURequestPage';
 import IOUBillPage from '../../../pages/iou/IOUBillPage';
 import IOUSendPage from '../../../pages/iou/IOUSendPage';
