+++ src/pages/workspace/GetAssistancePage.js
@@ -1,18 +1,18 @@
 import React, {Component} from 'react';
 import {View, ScrollView} from 'react-native';
-import ScreenWrapper from '../components/ScreenWrapper';
-import withLocalize, {withLocalizePropTypes} from '../components/withLocalize';
-import compose from '../libs/compose';
-import KeyboardAvoidingView from '../components/KeyboardAvoidingView';
-import HeaderWithCloseButton from '../components/HeaderWithCloseButton';
-import MenuItem from '../components/MenuItem';
-import Navigation from '../libs/Navigation/Navigation';
-import styles from '../styles/styles';
-import Text from '../components/Text';
-import RequestCallIcon from '../../assets/images/request-call.svg';
-import * as Expensicons from '../components/Icon/Expensicons';
-import * as Report from '../libs/actions/Report';
-import ROUTES from '../ROUTES';
+import ScreenWrapper from '../../components/ScreenWrapper';
+import withLocalize, {withLocalizePropTypes} from '../../components/withLocalize';
+import compose from '../../libs/compose';
+import KeyboardAvoidingView from '../../components/KeyboardAvoidingView';
+import HeaderWithCloseButton from '../../components/HeaderWithCloseButton';
+import MenuItem from '../../components/MenuItem';
+import Navigation from '../../libs/Navigation/Navigation';
+import styles from '../../styles/styles';
+import Text from '../../components/Text';
+import RequestCallIcon from '../../../assets/images/request-call.svg';
+import * as Expensicons from '../../components/Icon/Expensicons';
+import * as Report from '../../libs/actions/Report';
+import ROUTES from '../../ROUTES';
 
 const propTypes = {
     ...withLocalizePropTypes,
